Issue #5102: Refactoring of JavadocDetailNodeParser
