+++ config/findbugs-exclude.xml
@@ -93,7 +93,7 @@
     <Match>
         <!-- false-positive. Bugs reported even though casting is done only after verification using instanceof -->
         <Class name="com.puppycrawl.tools.checkstyle.JavadocDetailNodeParser"/>
-        <Method name="parseJavadocAsDetailNode"/>
+        <Method name="parseJavadoc"/>
         <Bug pattern="BC_UNCONFIRMED_CAST_OF_RETURN_VALUE"/>
     </Match>
     <Match>
